Fixing Implementation of the cart

+++ src/main/java/com/team29/backend/model/Product.java
@@ -17,7 +17,7 @@ public class Product {
     @GeneratedValue
     private Long id;
     private String name;
-    private Integer price;
+    private double price;
     private String image;
